% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm}
\alias{npm}
\alias{npm_run}
\alias{npx_run}
\title{NPM Utilities}
\usage{
npm_run(dir, ...)

npx_run(dir, ...)
}
\arguments{
\item{dir}{(dir) npm directory}

\item{...}{(arg) arguments to passed to \code{\link[processx:run]{processx::run}}}
}
\description{
Utiltiies to deal with \code{npm} and its attached technologies.
}
\section{Functions}{
\itemize{
\item \code{npm_run}: \verb{npm run} helper

\item \code{npx_run}: \verb{npx run} helper
}}

\seealso{
Other NPM Utilities: 
\code{\link{npm_install}()}
}
\concept{NPM Utilities}
