% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm-install.R
\name{npm_install}
\alias{npm_install}
\alias{npm_list}
\alias{npm_installed}
\title{NPM installation utilities}
\usage{
npm_install(dir, pkgs, ...)

npm_list(dir, pkgs, ...)

npm_installed(dir, pkgs, ...)
}
\arguments{
\item{dir}{(dir) npm directory}

\item{pkgs}{(arg) npm packages as character strings}

\item{...}{(arg) extra arguments to pass to \code{\link[processx:run]{processx::run}}}
}
\description{
NPM installation utilities
}
\section{Functions}{
\itemize{
\item \code{npm_install}: Install packages

\item \code{npm_list}: List packages in a directory

\item \code{npm_installed}: Checks whether certain npm packages are installed
}}

\seealso{
Other NPM Utilities: 
\code{\link{npm}}
}
\concept{NPM Utilities}
