% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind.R-package.R
\docType{package}
\name{tailwind.R-package}
\alias{tailwind.R}
\alias{tailwind.R-package}
\title{tailwind.R: Tailwind CSS for R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Tailwind CSS is a utility-first CSS framework that can be used
    to build any kind of design directly in the markup. This package allows
    R users to build Tailwind CSS pages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjpalanca/tailwind.R}
  \item Report bugs at \url{https://github.com/tjpalanca/tailwind.R/issues}
}

}
\author{
\strong{Maintainer}: Troy James Palanca \email{mail@tjpalanca.com} (\href{https://orcid.org/0000-0002-0610-9213}{ORCID})

}
\keyword{internal}
