% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind.R
\name{tailwind}
\alias{tailwind}
\title{Generate Tailwind CSS file}
\usage{
tailwind(
  config = tailwind_config(),
  input_file = NULL,
  output_file = "tailwind.css",
  working_dir = getwd()
)
}
\arguments{
\item{config}{(\code{\link[=turbo_config]{turbo_config()}}) instance governing this render}

\item{input_file}{(path) to an input CSS file. If NULL just renders the
default CSS files with default directives.}

\item{output_file}{(path) to save the rendered css file}

\item{working_dir}{(path) get the working directory}
}
\description{
Generate Tailwind CSS file
}
