% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind-assets.R
\name{tailwind_assets}
\alias{tailwind_assets}
\title{Tailwind CSS assets}
\usage{
tailwind_assets(
  input = pkg_inst("tailwind", "tailwind.css"),
  config = pkg_inst("tailwind", "tailwind.config.js"),
  output = pkg_inst("tailwind", "dist", "tailwind.css"),
  prod = FALSE,
  vers = NULL,
  build = !prod
)
}
\arguments{
\item{input}{(pth) path to the input css file with tailwind directives}

\item{config}{(pth) path to the config js file that dictates rendering}

\item{output}{(pth) path to the output css file that can be linked}

\item{prod}{(flg) whether this inclusion is in production}

\item{vers}{(flg) version of the tailwind dependency, defaults to the
hash digest of the current time}

\item{build}{(flg) whether to build the tailwind CSS file; default is FALSE
if \code{prod == TRUE}, and TRUE if \code{prod == FALSE}}
}
\description{
Include this in \code{{htmltools}} tags to include the rendered Tailwind CSS
file in the current page so that it can be used.
}
