% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind-config.R
\name{tailwind_config}
\alias{tailwind_config}
\alias{tailwind_purge}
\title{Tailwind CSS Configuration}
\usage{
tailwind_config(
  purge = tailwind_purge(),
  dark_mode = c("false", "media", "class"),
  theme = tailwind_theme(),
  variants = tailwind_variants(),
  plugins = tailwind_plugins()
)

tailwind_purge(enabled = TRUE, content = character(), safelist = character())
}
\arguments{
\item{purge}{(\code{\link[=tailwind_purge]{tailwind_purge()}}) instance}

\item{dark_mode}{(str) how dark mode is applied:
* false - don't apply
* media - ???
* class - ???}

\item{theme}{(\code{\link[=tailwind_theme]{tailwind_theme()}}) instance}

\item{variants}{(\code{\link[=tailwind_variants]{tailwind_variants()}}) instance}

\item{plugins}{(\code{\link[=tailwind_plugins]{tailwind_plugins()}}) instance}

\item{enabled}{(flg) whether or not the purge is enabled; disable for dev}

\item{content}{(chr) paths that will be scanned for the relevant classes}

\item{safelist}{(chr) classes accidentally removed from rendered CSS}
}
\description{
Tailwind CSS Configuration
}
\section{Functions}{
\itemize{
\item \code{tailwind_purge}: Purge settings (tree-shaking unused classes)
}}

