% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind-demo.R
\name{tailwind_demo}
\alias{tailwind_demo}
\alias{tailwind_demo_browse}
\title{Tailwind Demonstration}
\usage{
tailwind_demo()

tailwind_demo_browse()
}
\description{
This is a demonstration of an app that is built using the Tailwind CSS
JIT compiler.
}
