% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind-init.R
\name{tailwind_init}
\alias{tailwind_init}
\title{Initialize Tailwind Configs}
\usage{
tailwind_init(dir = pkg_inst("tailwind"), jit_mode = TRUE, overwrite = FALSE)
}
\arguments{
\item{dir}{(pth) where the config files will be stored.}

\item{jit_mode}{(flg) enable TailwindCSS jit mode}

\item{overwrite}{(flg) whether or not to overwrite}
}
\description{
The R package will store the PostCSS config, the Tailwind Config, and
the source directory within the project, while keeping the node machinery
outside of the package. This initializes those configs so that you can
run \code{tailwind_build()} easier.
}
\seealso{
Other Tailwind CSS: 
\code{\link{tailwind_assets}()},
\code{\link{tailwind_build}()},
\code{\link{tailwind_demo}()},
\code{\link{tailwind_install}()},
\code{\link{tailwind_run}()}
}
\concept{Tailwind CSS}
