% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind-run.R
\name{tailwind_run}
\alias{tailwind_run}
\title{Run Tailwind Commands}
\usage{
tailwind_run(...)
}
\arguments{
\item{...}{(arg) passed to \link[processx:run]{processx::run}}
}
\description{
After \code{tailwind_install()}, you can run tailwind commands arbitrarily using
\code{tailwind_run}.
}
