% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind-build.R
\name{tailwind_build}
\alias{tailwind_build}
\title{Build Tailwind CSS}
\usage{
tailwind_build(
  input = pkg_inst("tailwind", "tailwind.css"),
  config = pkg_inst("tailwind", "tailwind.config.js"),
  output = pkg_inst("tailwind", "dist", "tailwind.css"),
  prod = FALSE
)
}
\arguments{
\item{input}{(pth) path to the input css file with tailwind directives}

\item{config}{(pth) path to the config js file that dictates rendering}

\item{output}{(pth) path to the output css file that can be linked}

\item{prod}{(flg) to render production build vs development}
}
\description{
Builds the Tailwind CSS file based on the configuration in \code{dir}. It's
recommended that you use JIT mode when running \code{tailwind_install()} to make
this process faster.
}
\seealso{
Other Tailwind CSS: 
\code{\link{tailwind_assets}()},
\code{\link{tailwind_demo}()},
\code{\link{tailwind_init}()},
\code{\link{tailwind_install}()},
\code{\link{tailwind_run}()}
}
\concept{Tailwind CSS}
