% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind-install.R
\name{tailwind_install}
\alias{tailwind_install}
\alias{tailwind_installed}
\title{Install Tailwind CSS}
\usage{
tailwind_install()

tailwind_installed()
}
\description{
Installs the template tailwind css npm directory into a specific directory
that can then be operated by other functions in this package.

To configure the tailwind settings after installation, you should edit
the tailwind files themselves and refer to the
\href{https://tailwindcss.com/docs}{Tailwind Docs}.

Use \code{tailwind_installed()} to check a directory for an installation.
}
